requirements.txt
joblib
matplotlib
numpy
opencv_python
pandas
Pillow
streamlit
